♻️ refactor: extract EditablePhase to its own file
ScenarioEditorViewModel.swift had grown to 406 lines, tripping
swiftlint's file_length rule (limit 400). Split EditablePhase out —
it's a sibling editing-state type and matches the "file name = primary
type" naming convention.
